Fix type error

+++ firmware/coins.c
@@ -67,7 +67,7 @@ const CoinType *coinByName(const char *name)
 	return 0;
 }
 
-const CoinType *coinByAddressType(uint8_t address_type)
+const CoinType *coinByAddressType(uint32_t address_type)
 {
 	int i;
 	for (i = 0; i < COINS_COUNT; i++) {
